refactor validation code for modularity and sep of concerns
